+++ priv/templates/coh.install/controllers/coherence/session_controller.ex
@@ -184,25 +184,6 @@ defmodule <%= web_base %>.Coherence.SessionController do
   defp failed_login(conn, _user, _), do: put_flash(conn, :error, Messages.backend().incorrect_login_or_password(login_field: Config.login_field()))
 
   @doc """
-  Call back for the authentication plug.
-
-  Render the login form.
-  """
-  @spec login_callback(conn) :: conn
-  def login_callback(conn) do
-    conn =
-      if Map.get conn.private, :phoenix_layout do
-        conn
-      else
-        put_layout conn, Config.layout({Module.concat(Config.web_module, Coherence.LayoutView), :app})
-      end
-
-    conn
-    |> new(%{})
-    |> halt
-  end
-
-  @doc """
   Callback for the authenticate plug.
 
   Validate the rememberable cookie. If valid, generate a new token,
@@ -221,7 +202,7 @@ defmodule <%= web_base %>.Coherence.SessionController do
     case validate_login(id, series, token) do
       {:ok, rememberable} ->
         # Logger.debug "Valid login :ok"
-        Config.user_schema()
+        # Config.user_schema()
         id
         |> Schemas.get_user
         |> do_valid_login(conn, [id, rememberable, series, token], opts)